fix(NameAndParent): change length type from u8 to u32 for filename

+++ search-cache/src/slab_node.rs
@@ -9,9 +9,9 @@ use thin_vec::ThinVec;
 #[derive(Debug, Clone, Copy)]
 pub struct NameAndParent {
     ptr: *const u8,
-    // Length of the filename should not be larger than 256 bytes(macOS, Linux,
-    // Window, BSD) should be enough
-    len: u8,
+    // Length of the filename should not be larger than 256 chars(macOS, Linux,
+    // Window, BSD) should be enough.
+    len: u32,
     parent: crate::OptionSlabIndex,
 }
